chore: move ConfigOptions out of configurations
