+++ packages/middleware-bucket-endpoint/src/configurations.ts
@@ -1,4 +1,3 @@
-import { LoadedConfigSelectors } from "@aws-sdk/node-config-provider";
 import { Provider, RegionInfoProvider } from "@aws-sdk/types";
 
 export interface BucketEndpointInputConfig {
@@ -100,55 +99,3 @@ export function resolveBucketEndpointConfig<T>(
         : () => Promise.resolve(disableMultiregionAccessPoints),
   };
 }
-
-export const NODE_USE_ARN_REGION_ENV_NAME = "AWS_S3_USE_ARN_REGION";
-export const NODE_USE_ARN_REGION_INI_NAME = "s3_use_arn_region";
-
-export const NODE_DISABLE_MULTIREGION_ACCESS_POINT_ENV_NAME = "AWS_S3_DISABLE_MULTIREGION_ACCESS_POINTS";
-export const NODE_DISABLE_MULTIREGION_ACCESS_POINT_INI_NAME = "s3_disable_multiregion_access_points";
-
-/**
- * Config to load useArnRegion from environment variables and shared INI files
- *
- * @api private
- */
-export const NODE_USE_ARN_REGION_CONFIG_OPTIONS: LoadedConfigSelectors<boolean> = {
-  environmentVariableSelector: (env: NodeJS.ProcessEnv) => {
-    if (!Object.prototype.hasOwnProperty.call(env, NODE_USE_ARN_REGION_ENV_NAME)) return undefined;
-    if (env[NODE_USE_ARN_REGION_ENV_NAME] === "true") return true;
-    if (env[NODE_USE_ARN_REGION_ENV_NAME] === "false") return false;
-    throw new Error(
-      `Cannot load env ${NODE_USE_ARN_REGION_ENV_NAME}. Expected "true" or "false", got ${env[NODE_USE_ARN_REGION_ENV_NAME]}.`
-    );
-  },
-  configFileSelector: (profile) => {
-    if (!Object.prototype.hasOwnProperty.call(profile, NODE_USE_ARN_REGION_INI_NAME)) return undefined;
-    if (profile[NODE_USE_ARN_REGION_INI_NAME] === "true") return true;
-    if (profile[NODE_USE_ARN_REGION_INI_NAME] === "false") return false;
-    throw new Error(
-      `Cannot load shared config entry ${NODE_USE_ARN_REGION_INI_NAME}. Expected "true" or "false", got ${profile[NODE_USE_ARN_REGION_INI_NAME]}.`
-    );
-  },
-  default: false,
-};
-
-export const NODE_DISABLE_MULTIREGION_ACCESS_POINT_CONFIG_OPTIONS: LoadedConfigSelectors<boolean> = {
-  environmentVariableSelector: (env: NodeJS.ProcessEnv) => {
-    if (!Object.prototype.hasOwnProperty.call(env, NODE_DISABLE_MULTIREGION_ACCESS_POINT_ENV_NAME)) return undefined;
-    if (env[NODE_DISABLE_MULTIREGION_ACCESS_POINT_ENV_NAME] === "true") return true;
-    if (env[NODE_DISABLE_MULTIREGION_ACCESS_POINT_ENV_NAME] === "false") return false;
-    throw new Error(
-      `Cannot load env ${NODE_DISABLE_MULTIREGION_ACCESS_POINT_ENV_NAME}. Expected "true" or "false", got ${env[NODE_DISABLE_MULTIREGION_ACCESS_POINT_ENV_NAME]}.`
-    );
-  },
-  configFileSelector: (profile) => {
-    if (!Object.prototype.hasOwnProperty.call(profile, NODE_DISABLE_MULTIREGION_ACCESS_POINT_INI_NAME))
-      return undefined;
-    if (profile[NODE_DISABLE_MULTIREGION_ACCESS_POINT_INI_NAME] === "true") return true;
-    if (profile[NODE_DISABLE_MULTIREGION_ACCESS_POINT_INI_NAME] === "false") return false;
-    throw new Error(
-      `Cannot load shared config entry ${NODE_DISABLE_MULTIREGION_ACCESS_POINT_INI_NAME}. Expected "true" or "false", got ${profile[NODE_DISABLE_MULTIREGION_ACCESS_POINT_INI_NAME]}.`
-    );
-  },
-  default: false,
-};
